requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.0
pytesseract==0.3.10
pillow==10.1.0
langchain==0.0.346
chromadb==0.4.18
sentence-transformers==2.2.2
pdf2image==1.16.3
pymongo==4.5.0
redis==5.0.1
celery==5.3.1
python-dotenv==1.0.0
python-magic==0.4.27
httpx==0.25.1
google-genai